+++ index.d.ts
@@ -20,6 +20,7 @@ export interface Config {
   readonly replace?: string;
   readonly redirect?: (path: string) => string;
   readonly fetch?: {
+    readonly headers?: Headers;
     readonly timeout?: number;
     readonly wait?: number;
   };
@@ -38,7 +39,7 @@ export interface Config {
 }
 
 export interface Sequence<a, b, c, d> {
-  readonly fetch: (result: void, request: { path: string; method: string; body: FormData | null; }) => Promise<a>;
+  readonly fetch: (result: void, request: { path: string; method: string; headers: Headers; body: FormData | null; }) => Promise<a>;
   readonly unload: (result: a, response: { url: string; header: (name: string) => string | null; document: Document; }) => Promise<b>;
   readonly content: (result: b, areas: HTMLElement[]) => Promise<c>;
   readonly ready: (result: c) => Promise<d>;